% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelMiss.R
\name{DelMiss}
\alias{DelMiss}
\title{Delete variables that contain missing values}
\usage{
DelMiss(eSet)
}
\arguments{
\item{eSet}{A R6 class object}
}
\value{
A R6 class object
}
\description{
Delete variables that contain missing values
}
\examples{
eSet <- DelMiss(eSet = eSet)
}
