% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelNearZeroVar.R
\name{DelNearZeroVar}
\alias{DelNearZeroVar}
\title{Title}
\usage{
DelNearZeroVar(eSet)
}
\arguments{
\item{eSet}{A R6 class object}
}
\value{
A R6 class object
}
\description{
Title
}
\examples{
eSet <- DelNearZeroVar(eSet = eSet)
}
