% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IniMedt.R
\name{IniMedt}
\alias{IniMedt}
\title{Initialize for ExpoMediation}
\usage{
IniMedt(
  PID = "auto",
  FileDirIn = "default",
  FileDirOut = "default",
  Module = NULL
)
}
\arguments{
\item{PID}{A unique identification for user}

\item{FileDirIn}{The physical path that contains the input files}

\item{FileDirOut}{The output path that stores the result files.}

\item{Module}{Mediation}
}
\value{
eSet, a R6 class object
}
\description{
Initialize for ExpoMediation
}
\examples{
eSet <- IniMedt(PID = "R1",
FileDirIn = "default",
FileDirOut = "default",
Module = "Mediation")
}
