% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDataMedt.R
\name{LoadDataMedt}
\alias{LoadDataMedt}
\title{Load data for mediation analyses}
\usage{
LoadDataMedt(
  eSet,
  UseExample = "default",
  FileDirExpo = NULL,
  FileDirVoca = NULL
)
}
\arguments{
\item{eSet}{A R6 class containing all information that users have provided}

\item{UseExample}{True/False, whether use example data}

\item{FileDirExpo}{The name of data file to be uploaded}

\item{FileDirVoca}{The name of vocabulary file to be uploaded}
}
\value{
eSet, a R6 class object
}
\description{
Load data for mediation analyses
}
\examples{
eSet <- LoadData(eSet = eSet,
UseExample = "example1",
FileDirExpo = NULL,
FileDirVoca = NULL)
}
