% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedtImptM.R
\name{MedtImptM}
\alias{MedtImptM}
\title{Calculate the importance of each mediator}
\usage{
MedtImptM(
  eSet = eSet,
  VarsY = NULL,
  VarsX = eSet$Expo$Voca \%>\% dplyr::filter(GroupName == "chemical") \%>\%
    dplyr::select(SerialNo) \%>\% as.matrix() \%>\% as.vector(),
  VarsC = eSet$Expo$Voca \%>\% dplyr::filter(stringr::str_detect(SerialNo, "^C")) \%>\%
    dplyr::select(SerialNo) \%>\% as.matrix() \%>\% as.vector()
)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{VarsY}{A character indicating the name of outcome}

\item{VarsX}{A character vector indicating the name of exposures}

\item{VarsC}{A character vector indicating the name of covariates}
}
\value{
eSet, A R6 class object
}
\description{
Calculate the importance of each mediator
}
\examples{
eSet <- MedtImptM(eSet = eSet,
                  VarsY = "Y1",
                  VarsX = c("X1", "X2", "X3"),
                  VarsC = c("C1.2", "C1.3", "C2", "C3", "C4", "C5.2", "C6.2", "C6.3", "C6.4"))
}
