% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransSubgroup.R
\name{MedtMlist}
\alias{MedtMlist}
\title{Divide mediators into separate groups according to the dictionary dataset}
\usage{
MedtMlist(eSet, voca = eSet$Expo$Voca)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{voca}{Vocabulary dataset containing variable information}
}
\value{
eSet, A R6 class object
}
\description{
Divide mediators into separate groups according to the dictionary dataset
}
\examples{
eSet <- MedtMlist(eSet = eSet)
}
