% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedtPairWise.R
\name{MedtPairWise}
\alias{MedtPairWise}
\title{Implementing pair-wise mediation analyses for each exposure-mediator combination}
\usage{
MedtPairWise(eSet, VarsY, VarsX, VarsM, VarsC, Iter)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{VarsY}{A character indicating the name of outcome}

\item{VarsX}{A character vector indicating the name of exposures}

\item{VarsM}{A character vector indicating the name of mediator}

\item{VarsC}{A character vector indicating the name of covariates}

\item{Iter}{The number of iterations of mediation modeling}
}
\value{
eSet, A R6 class object
}
\description{
Implementing pair-wise mediation analyses for each exposure-mediator combination
}
\examples{
eSet <- MedtPairWise(eSet = eSet,
                     VarsY = "Y1",
                     VarsX = c("X1","X2","X3", "X10", "X11", "X12", "X19", "X20", "X21"),
                     VarsM = c("M1","M2","M3","M4","M5","M8","M8"),
                     VarsC = c("C1.2", "C1.3", "C2", "C3", "C4", "C5.2", "C6.2", "C6.3", "C6.4"),
                     Iter = 500)
}
