% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedtRedM.R
\name{MedtRedM}
\alias{MedtRedM}
\title{Mediator dimension reduction: extract the first principal direction of mediator}
\usage{
MedtRedM(
  eSet = eSet,
  VarsY = "Y1",
  VarsX = NULL,
  VarsC = eSet$Expo$Voca \%>\% dplyr::filter(str_detect(SerialNo, "^C")) \%>\%
    dplyr::select(SerialNo) \%>\% as.matrix() \%>\% as.vector(),
  Method = "mean",
  Iter = 500
)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{VarsY}{A character indicating the name of outcome}

\item{VarsX}{A character vector indicating the name of exposures}

\item{VarsC}{A character vector indicating the name of covariates}

\item{Method}{Dimension reduction method, available options are "mean"(default) and "PDM1"}

\item{Iter}{Iterations of mediation models,default is 500}
}
\value{
eSet, A R6 class object
}
\description{
Mediator dimension reduction: extract the first principal direction of mediator
}
\examples{
eSet <- MedtRedM(eSet = eSet,
                 VarsY = "Y1",
                 VarsX = c("X1", "X2", "X3"),
                 VarsC = c("C1.2", "C1.3", "C2", "C3", "C4", "C5.2", "C6.2", "C6.3", "C6.4"),
                 Method = "mean",
                 Iter = 500)
}
