% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedtRedXM.R
\name{MedtRedXM}
\alias{MedtRedXM}
\title{Exposure and mediator dimension reduction}
\usage{
MedtRedXM(eSet, Method = "mean", Iter = 500)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{Method}{Method used for mediator reduction wth all exposures, available options are "mean"(default) and "pdm1"}

\item{Iter}{Iterations of mediation models,default is 500}
}
\value{
eSet, A R6 class object
}
\description{
Exposure and mediator dimension reduction
}
\examples{
eSet <- MedtRedXM(eSet = eSet,
                  Method = "mean",
                  Iter = 500)
}
