% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransDistr.R
\name{TransDistr}
\alias{TransDistr}
\title{Transform distribution}
\usage{
TransDistr(eSet, Vars, Method)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{Vars}{Variables to be transformed}

\item{Method}{Transform distribution methods, including "ln", "log2" and "log10"}
}
\value{
eSet, A R6 class object
}
\description{
Transform distribution
}
\examples{
eSet <- TransDistr(eSet = eSet,
                   Vars = c("X1", "X2", "X3"),
                   Method = "log10")
}
