% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransDummy.R
\name{TransDummy}
\alias{TransDummy}
\title{Transform (factor) variables into dummy variables}
\usage{
TransDummy(eSet, Vars = "default")
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{Vars}{Variables (factor) to be transformed}
}
\value{
eSet, A R6 class object
}
\description{
Transform (factor) variables into dummy variables
}
\examples{
eSet <- TransDummy(eSet = eSet,
                   Vars = c("C1", "C5", "C6"))
}
