% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransImput.R
\name{TransImput}
\alias{TransImput}
\title{Transform for data imputation}
\usage{
TransImput(eSet, Group = T, Vars, Method = NULL)
}
\arguments{
\item{eSet}{A R6 class object containning the dataset}

\item{Group}{True/False, indicating whether execute imputation by group variable}

\item{Vars}{Variables to be imputed}

\item{Method}{Imputation methods, available methods include "cart", "lod"}
}
\value{
eSet, a R6 class object
}
\description{
Transform for data imputation
}
\examples{
eSet <- TransImput(eSet = eSet,
                   Group = F,
                   Vars = c("X3","M22","M29"),
                   Method = "lod")
}
