% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransType.R
\name{TransType}
\alias{TransType}
\title{Transform Data type}
\usage{
TransType(eSet, Vars, TypeTo)
}
\arguments{
\item{eSet}{eSet, A R6 class object}

\item{Vars}{Variables to be transformed}

\item{TypeTo}{Transform methods, available methods include "integer", "numeric", "character", "factor", "logical" and "date"}
}
\value{
eSet, A R6 class object
}
\description{
Transform Data type
}
\examples{
eSet <- TransType(eSet = eSet,
                  Vars = c("C1", "C5", "C6"),
                  TypeTo = "factor")
}
