% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMedtPair.R
\name{VizMedtPair}
\alias{VizMedtPair}
\title{Visualize pairwise mediation modelling result}
\usage{
VizMedtPair(eSet = eSet)
}
\arguments{
\item{eSet}{eSet, A R6 class object}
}
\value{
eSet, A R6 class object
}
\description{
Visualize pairwise mediation modelling result
}
\examples{
eSet <- VizMedtPair(eSet = eSet)
}
