% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMedtRedXM.R
\name{VizMedtRedXM}
\alias{VizMedtRedXM}
\title{Visualize MedtRedXM mediation modelling result}
\usage{
VizMedtRedXM(eSet = eSet)
}
\arguments{
\item{eSet}{eSet, A R6 class object}
}
\value{
eSet, A R6 class object
}
\description{
Visualize MedtRedXM mediation modelling result
}
\examples{
eSet <- VizMedtRedXM(eSet = eSet)
}
